implement hexagonal architecture
