+++ .gitignore
@@ -2,7 +2,9 @@
 __pycache__/
 *.py[cod]
 *$py.class
-
+gen
+*.gen.go
+./
 ./.idea
 /.idea
 # C extensions
