+++ services/app-auth/internal/adapters/grpc/user_grpc_repository.go
@@ -1,0 +1,1 @@
+package grpc
