+++ services/app-auth/internal/core/entity/user.go
@@ -1,4 +1,4 @@
-package models
+package entity
 
 type User struct {
 	Name        string
