+++ services/app-auth/internal/core/repository/user_grcp_repository.go
@@ -1,1 +1,1 @@
-package repositories
+package repository
