+++ services/app-auth/internal/core/usecase/auth.go
@@ -1,0 +1,1 @@
+package usecase
