+++ services/app-auth/internal/platform/database/grpc/app_db_client.go
@@ -1,1 +1,18 @@
 package grpc
+
+import (
+	"context"
+	"github.com/AppsLab-KE/backend-everyshilling/services/app-authentication/internal/core/entity"
+)
+
+type Client struct {
+	client *string
+}
+
+func (c Client) Create(ctx context.Context, data entity.User) error {
+	return nil
+}
+
+func NewClient() *Client {
+	return &Client{}
+}
