+++ services/app-auth/main.go
@@ -1,7 +1,4 @@
 package main
 
-import "github.com/AppsLab-KE/backend-everyshilling/services/app-authentication/cmd"
-
 func main() {
-	cmd.Execute()
 }
