+++ services/app-otp/verification/admin.py
@@ -1,3 +1,3 @@
 from django.contrib import admin
 
-# Register your models here.
+# Register your entity here.
