+++ services/app-otp/verification/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 
 class VerificationConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
+    default_auto_field = 'django.db.entity.BigAutoField'
     name = 'verification'
