+++ services/app-otp/verification/models.py
@@ -1,3 +1,3 @@
 from django.db import models
 
-# Create your models here.
+# Create your entity here.
